backend/requirements.txt
# Core backend dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-dotenv>=1.0.0

# AI/ML Libraries
openai>=1.3.0
langchain-gradient>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.60
pydantic>=2.5.0
typing-extensions>=4.8.0

# Search functionality - SERP API
google-search-results>=2.4.0

# Frontend/UI (optional - for Streamlit interface)
streamlit>=1.28.0
pydeck>=0.8.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Additional utilities
annotated-types>=0.6.0

# Note: If you encounter SERP API import issues, you may also need:
# serpapi>=0.1.0